+++ changelog.md
@@ -1,3 +1,5 @@
+Version 1.10.9
+
 - Aim-assisted turret weapons will not attack when wildly off target.
 - Aim assist no longer interferes with movement systems. Reported by Rubenlee123.
 - Invictus and other ships controller by custom ship AI aggressively follow the eliminate order. Issue reported by mora.